Create ESP-32 Arduino Code
This is the code that we used for our door touch sensor. It takes a reading when someone touches the handle and tries to send the value to our database. If it can't connect to the Wi-Fi, it will store the value in an array and try again later. When it is not taking readings, it is in a sleep state which reduces power consumption.